requirements.txt
sentencepiece
fastapi
uvicorn
torch
transformers
sacrebleu
requests
pytest
gitpython


